requirements.txt
streamlit==1.32.2
easyocr==1.7.1
gTTS==2.5.1
googletrans==4.0.0-rc1
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4